Add Azure deployment setup and ARM template
Added a new publish profile for web deployment in CRUD application 2.csproj, and introduced a new Azure Resource Manager (ARM) template (`profile.arm.json`) to define infrastructure for Azure resources. This includes setup for an Azure App Service on Windows, with parameters and variables for resource management, and specifies an Incremental deployment mode. The ARM template outlines the creation of a resource group and an Azure App Service with a .NET Core stack, under a Standard tier App Service Plan.